+++ src/components/ArbiterRecall.tsx
@@ -1,8 +1,12 @@
 import * as React from 'react';
-import { loadModule, setupModule, defaultFetchDependency } from '../utils';
+import { setupModules, loadModules } from '../utils';
 import { Module, ArbiterDisplay, ArbiterOptions } from '../types';
 
 export interface ArbiterRecallProps<TApi> extends ArbiterOptions<TApi> {
+  /**
+   * The component to render when the modules have been loaded
+   * (node instance) or with the state of the recall (function).
+   */
   children?: React.ReactChild | ArbiterDisplay<TApi>;
 }
 
@@ -12,6 +16,9 @@ export interface ArbiterState<TApi> {
   modules: Array<Module<TApi>>;
 }
 
+/**
+ * Represents an arbiter recall component to load extension components.
+ */
 export class ArbiterRecall<TApi> extends React.Component<ArbiterRecallProps<TApi>, ArbiterState<TApi>> {
   private mounted = false;
 
@@ -23,37 +30,24 @@ export class ArbiterRecall<TApi> extends React.Component<ArbiterRecallProps<TApi
     };
   }
 
-  private finish(modules: Array<Module<TApi>>, error: any) {
-    const { createApi } = this.props;
-
-    if (typeof createApi === 'function') {
-      for (const app of modules) {
-        const api = createApi(app);
-        setupModule(app, api);
-      }
-    } else {
-      console.warn('Invalid `createApi` prop. Skipping module installation.');
-    }
+  private finish(error: any, newModules: Array<Module<TApi>>) {
+    const { createApi, modules: oldModules = [] } = this.props;
 
     this.setState({
-      loaded: true,
       error,
-      modules,
+      loaded: true,
+      modules: setupModules(createApi, [...oldModules, ...newModules]),
     });
   }
 
   componentDidMount() {
-    const { getModules, dependencies = {}, fetchDependency = defaultFetchDependency, modules = [] } = this.props;
+    const { getModules, dependencies, fetchDependency } = this.props;
     this.mounted = true;
 
-    if (typeof getModules === 'function') {
-      Promise.resolve(getModules())
-        .then(moduleData => Promise.all(moduleData.map(m => loadModule<TApi>(m, fetchDependency, dependencies))))
-        .then(newModules => this.mounted && this.finish([...newModules, ...modules], undefined))
-        .catch(error => this.mounted && this.finish([], error));
-    } else {
-      console.error('Could not get the modules. Provide a valid `getModules` function as prop.');
-    }
+    loadModules(getModules, fetchDependency, dependencies).then(
+      modules => this.mounted && this.finish(undefined, modules),
+      error => this.mounted && this.finish(error, []),
+    );
   }
 
   componentWillUnmount() {