+++ src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './aggregate';
 export * from './dependency';
 export * from './fetch';
 export * from './load';