+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # React Arbiter Changelog
 
+## 0.2.0
+
+- Added documentation comments
+- Additional options for `wrapComponent` and `wrapElement`
+- Introduced `setupModules` and `loadModules` helper
+
 ## 0.1.0
 
 - Initial Release